refactor: rename Telemetry module into `Telemetry`
