+++ src/commands/DebuggerCommands.ts
@@ -10,7 +10,7 @@ import {TransactionProvider} from '@/debugAdapter/transaction/transactionProvide
 import {Web3Wrapper} from '@/debugAdapter/web3Wrapper';
 import {getTruffleWorkspace, getPathByPlatform} from '@/helpers/workspace';
 import {showInputBox, showQuickPick} from '@/helpers/userInteraction';
-import {Telemetry} from '@/TelemetryClient';
+import {Telemetry} from '@/Telemetry';
 import {DebuggerTypes} from '@/debugAdapter/models/debuggerTypes';
 
 const TX_REGEX = /^(?:0x)?[0-9a-fA-F]{64}$/;
