+++ src/helpers/TruffleConfiguration.ts
@@ -1,6 +1,6 @@
 import {Constants} from '@/Constants';
 import {MnemonicRepository} from '@/services/MnemonicRepository';
-import {Telemetry} from '@/TelemetryClient';
+import {Telemetry} from '@/Telemetry';
 import * as acorn from 'acorn';
 import {Node} from 'acorn';
 import * as walk from 'acorn-walk';
