+++ src/helpers/userInteraction.ts
@@ -5,7 +5,7 @@ import fs from 'fs';
 import {InputBoxOptions, ProgressLocation, QuickPickItem, QuickPickOptions, Uri, window, workspace} from 'vscode';
 import {Constants, NotificationOptions} from '@/Constants';
 import {CancellationEvent} from '@/Models/CancellationEvent';
-import {Telemetry} from '@/TelemetryClient';
+import {Telemetry} from '@/Telemetry';
 import {DialogResultValidator} from '@/validators/DialogResultValidator';
 
 export async function showInputBox(options: InputBoxOptions): Promise<string> {
