+++ src/pages/Changelog.ts
@@ -5,7 +5,7 @@ import fs from 'fs-extra';
 import semver from 'semver';
 import {commands, ExtensionContext} from 'vscode';
 import {Constants} from '../Constants';
-import {Telemetry} from '../TelemetryClient';
+import {Telemetry} from '../Telemetry';
 import {BasicWebView, IWebViewConfig} from './BasicWebView';
 
 export class ChangelogPage extends BasicWebView {
