+++ src/resourceExplorers/GenericResourceExplorer.ts
@@ -6,7 +6,7 @@ import {showInputBox} from '@/helpers/userInteraction';
 import {GenericNetworkNode} from '@/Models/TreeItems/GenericNetworkNode';
 import {GenericProject} from '@/Models/TreeItems/GenericProject';
 import {GenericService} from '@/services/generic/GenericService';
-import {Telemetry} from '@/TelemetryClient';
+import {Telemetry} from '@/Telemetry';
 import {DialogResultValidator} from '@/validators/DialogResultValidator';
 import {UrlValidator} from '@/validators/UrlValidator';
