+++ src/resourceExplorers/InfuraResourceExplorer.ts
@@ -10,7 +10,7 @@ import {InfuraProject} from '@/Models/TreeItems/InfuraProject';
 import {InfuraLayer} from '@/Models/TreeItems/InfuraLayer';
 import {IInfuraEndpointDto, IInfuraProjectDto, IInfuraProjectQuickPick} from '@/services/infuraService/InfuraDto';
 import {InfuraServiceClient} from '@/services/infuraService/InfuraServiceClient';
-import {Telemetry} from '@/TelemetryClient';
+import {Telemetry} from '@/Telemetry';
 
 export class InfuraResourceExplorer {
   public async createProject(existingProjects: string[] = []): Promise<InfuraProject> {
