+++ src/resourceExplorers/LocalResourceExplorer.ts
@@ -6,7 +6,7 @@ import {showInputBox} from '@/helpers/userInteraction';
 import {LocalProject, TLocalProjectOptions} from '@/Models/TreeItems/LocalProject';
 import {LocalNetworkNode} from '@/Models/TreeItems/LocalNetworkNode';
 import {GanacheService} from '@/services/ganache/GanacheService';
-import {Telemetry} from '@/TelemetryClient';
+import {Telemetry} from '@/Telemetry';
 import {DialogResultValidator} from '@/validators/DialogResultValidator';
 import {UrlValidator} from '@/validators/UrlValidator';
