+++ src/services/contract/ContractDB.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 
 import {EventEmitter} from 'events';
-import {Telemetry} from '../../TelemetryClient';
+import {Telemetry} from '../../Telemetry';
 import {AbstractAdapter} from './AbstractAdapter';
 import {ContractInstance} from './ContractInstance';
 import {InMemoryAdapter} from './InMemoryAdapter';
