+++ src/services/contract/InMemoryAdapter.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 
 import {Output, OutputLabel} from '@/Output';
-import {Telemetry} from '../../TelemetryClient';
+import {Telemetry} from '../../Telemetry';
 import {NetworkService} from '../NetworkService';
 import {AbstractAdapter} from './AbstractAdapter';
 import {Contract} from './Contract';
