+++ src/services/dashboard/DashboardServiceClient.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 
 import {Constants} from '../../Constants';
-import {Telemetry} from '../../TelemetryClient';
+import {Telemetry} from '../../Telemetry';
 import {HttpService} from '../HttpService';
 
 export async function isDashboardRunning(port: number | string): Promise<boolean> {
