+++ src/services/generic/GenericService.ts
@@ -5,7 +5,7 @@ import type {ChildProcess} from 'child_process';
 import type {OutputChannel} from 'vscode';
 import {Constants} from '../../Constants';
 import * as shell from '@/helpers/shell';
-import {Telemetry} from '../../TelemetryClient';
+import {Telemetry} from '../../Telemetry';
 import {UrlValidator} from '../../validators/UrlValidator';
 import {isGenericServer, getWeb3ClientVersion} from './GenericServiceClient';
