+++ src/services/generic/GenericServiceClient.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 
 import {Constants} from '../../Constants';
-import {Telemetry} from '../../TelemetryClient';
+import {Telemetry} from '../../Telemetry';
 import {HttpService} from '../HttpService';
 
 export async function isGenericServer(port: number | string): Promise<boolean> {
