+++ src/services/infuraService/codeFlowLogin.ts
@@ -9,7 +9,7 @@ import requestPromise from 'request-promise';
 import url from 'url';
 import {env, Uri} from 'vscode';
 import {Constants} from '../../Constants';
-import {Telemetry} from '../../TelemetryClient';
+import {Telemetry} from '../../Telemetry';
 
 interface IDeferred<T> {
   resolve: (result: T | Promise<T>) => void;
